Use system CA certs bundle if it exists
libcloud seems to take the first CA_CERT_PATH that it finds.  This commit
appends data/cacert.pem from the repo instead of inserting it.  This let's us
default to the system CA certs and only use the CA bundle from this repo
if we don't find one of the libcloud defaults.  This allows a deployer
to manage internal CA certs using native distro tools (e.g.
update-ca-certificates).

Closes: https://github.com/racker/rackspace-monitoring-cli/issues/85